backend/app/requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
motor==3.1.2
numpy==1.24.3
pandas==2.0.2
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
